# backend/docs/openapi.json
{
    "components": {
        "schemas": {
...
            "quiz.AttemptAnswers": {
                "properties": {
                    "answers": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "example": {
                            "0": 1
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "quiz.AttemptResponse": {
                "properties": {
                    "answers": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object"
                    },
                    "feedback": {
                        "items": {
                            "type": "boolean"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "id": {
                        "example": 1,
                        "type": "integer"
                    },
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/quiz.StudentQuestion"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "score": {
                        "example": 100,
                        "type": "integer"
                    },
                    "status": {
                        "example": "submitted",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "quiz.CreateQuizRequest": {
                "properties": {
                    "attempts_allowed": {
                        "example": 3,
                        "type": "integer"
                    },
                    "feedback_policy": {
                        "example": "on_submit",
                        "type": "string"
                    },
                    "resource_id": {
                        "example": 7,
                        "type": "integer"
                    },
                    "time_limit_sec": {
                        "example": 600,
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "quiz.QuizDetailResponse": {
                "properties": {
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/quiz.Question"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "quiz": {
                        "$ref": "#/components/schemas/quiz.Quiz"
                    }
                },
                "type": "object"
            },
            "quiz.QuizEnvelope": {
                "properties": {
                    "quiz": {
                        "$ref": "#/components/schemas/quiz.Quiz"
                    }
                },
                "type": "object"
            },
...
            "utils.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "Se requieren username, email y password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "utils.MessageResponse": {
                "properties": {
                    "message": {
                        "example": "Operación exitosa",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Token JWT de sesión. Formato: `Bearer \u003ctoken\u003e`.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "API REST del monolito modular: identidad y sesiones, administración y auditoría, autoría de cursos versionados y carga multimedia con procesamiento asíncrono a HLS.\nAutenticación con JWT revocable: iniciar sesión en /api/v1/auth/login y enviar `Authorization: Bearer \u003ctoken\u003e`.",
        "title": "Plataforma MOOC - API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/v1/attempts/{id}": {
            "get": {
                "description": "Score y feedback solo tras el envío, según la política del quiz.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del intento",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.AttemptResponse"
                                }
                            }
                        },
                        "description": "Intento"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Ver intento",
                "tags": [
                    "Quizzes"
                ]
            },
            "put": {
                "description": "Solo sobre intentos en borrador vigente; valida rangos en servidor.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del intento",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/quiz.AttemptAnswers",
                                "summary": "request",
                                "description": "Respuestas (posición → opción)"
                            }
                        }
                    },
                    "description": "Respuestas (posición → opción)",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.AttemptResponse"
                                }
                            }
                        },
                        "description": "Intento actualizado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Respuesta fuera de rango"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Intento cerrado o expirado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Guardado parcial",
                "tags": [
                    "Quizzes"
                ]
            }
        },
        "/api/v1/attempts/{id}/submit": {
            "post": {
                "description": "Calcula la nota en servidor desde el snapshot; nunca confía en el cliente.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del intento",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.AttemptResponse"
                                }
                            }
                        },
                        "description": "Calificado"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Intento cerrado o expirado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Enviar intento",
                "tags": [
                    "Quizzes"
                ]
            }
        },
...
        "/api/v1/quizzes": {
            "post": {
                "description": "Define intentos, límite de tiempo y retroalimentación sobre un recurso quiz del borrador.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/quiz.CreateQuizRequest",
                                "summary": "request",
                                "description": "Reglas del quiz"
                            }
                        }
                    },
                    "description": "Reglas del quiz",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.QuizEnvelope"
                                }
                            }
                        },
                        "description": "Quiz creado"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Datos inválidos"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Sin permiso"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Versión inmutable"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Crear quiz",
                "tags": [
                    "Quizzes"
                ]
            }
        },
        "/api/v1/quizzes/{id}": {
            "delete": {
                "description": "Borra el quiz con sus preguntas e intentos. Solo en versión borrador de curso no publicado.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del quiz",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.MessageResponse"
                                }
                            }
                        },
                        "description": "Quiz eliminado"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Sin permiso"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Versión inmutable o curso publicado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Eliminar quiz",
                "tags": [
                    "Quizzes"
                ]
            },
            "get": {
                "description": "Incluye correct_index; solo professor/admin propietarios.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del quiz",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.QuizDetailResponse"
                                }
                            }
                        },
                        "description": "Quiz con preguntas"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Sin permiso"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Ver quiz (autoría)",
                "tags": [
                    "Quizzes"
                ]
            }
        },
        "/api/v1/quizzes/{id}/attempts": {
            "post": {
                "description": "Congela un snapshot de preguntas; reintento con la misma clave retorna el mismo intento.",
                "parameters": [
                    {
                        "description": "Bearer \u003ctoken\u003e",
                        "in": "header",
                        "name": "Authorization",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Clave de idempotencia",
                        "in": "header",
                        "name": "Idempotency-Key",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "ID del quiz",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/quiz.AttemptResponse"
                                }
                            }
                        },
                        "description": "Intento"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "Sin derecho o intentos agotados"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/utils.ErrorResponse"
                                }
                            }
                        },
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Iniciar intento",
                "tags": [
                    "Quizzes"
                ]
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "localhost:8080"
        }
    ]
}